+++ s/base/helpers/apply_theme.ts
@@ -2,8 +2,8 @@
 import {CSSResultGroup} from "lit"
 
 import {mixinCss} from "../mixins/css.js"
+import {Elements} from "../utils/types.js"
 import {obtool} from "../../tools/obtool.js"
-import {Elements} from "../utils/elements.js"
 
 export const apply_theme = (theme: CSSResultGroup) => {
 	return <E extends Elements>(elements: E) => obtool(elements).map(