+++ s/base/helpers/cue_reactivity.ts
@@ -1,8 +1,8 @@
 
+import {Elements} from "../utils/types.js"
 import {mixinCues} from "../mixins/cues.js"
 import {CueGroup} from "../../cues/group.js"
 import {obtool} from "../../tools/obtool.js"
-import {Elements} from "../utils/elements.js"
 
 export function cue_reactivity(cues: CueGroup) {
 	return <E extends Elements>(elements: E) => obtool(elements).map(