+++ s/base/helpers/flatstate_reactivity.ts
@@ -1,8 +1,8 @@
 
-import {obtool} from "../../tools/obtool.js"
-import {Elements} from "../utils/elements.js"
-import {mixinFlatstate} from "../mixins/flatstate.js"
+import {Elements} from "../utils/types.js"
 import {Flat} from "../../flatstate/flat.js"
+import {obtool} from "../../tools/obtool.js"
+import {mixinFlatstate} from "../mixins/flatstate.js"
 
 export function flatstate_reactivity(flat: Flat) {
 	return <E extends Elements>(elements: E) => obtool(elements).map(