+++ s/base/helpers/snap_reactivity.ts
@@ -1,9 +1,9 @@
 
 import type {Snapstate} from "@chasemoskal/snapstate"
 
+import {Elements} from "../utils/types.js"
 import {mixinSnap} from "../mixins/snap.js"
 import {obtool} from "../../tools/obtool.js"
-import {Elements} from "../utils/elements.js"
 
 export function snap_reactivity(snaps: Snapstate<any>[]) {
 	return <E extends Elements>(elements: E) => obtool(elements).map(