+++ s/demo/components/local-counter.ts
@@ -1,22 +1,28 @@
 
 import {html} from "lit"
+
+import {Context} from "../context.js"
 import {common_styles} from "../common-styles.js"
 import {QuickElement} from "../../quick/element.js"
+import {attributes} from "../../base/addons/attributes.js"
 
-export const LocalCounter = () => class extends QuickElement {
-
+export const LocalCounter = (_: Context) => class extends QuickElement {
 	static styles = common_styles
 
-	count = this.cues.create(0)
+	attr = attributes<{count: string}>(this as QuickElement)
+
+	get count() {
+		return this.attr.number.count ?? 0
+	}
 
 	#increment = () => {
-		this.count.value += 1
+		this.attr.number.count = this.count + 1
 	}
 
 	render() {
 		return html`
 			<p class=tag>&lt;local-counter&gt;</p>
-			<p class=count>${this.count.value}</p>
+			<p class=count>${this.attr.number.count}</p>
 			<button @click=${this.#increment}>increment</button>
 		`
 	}